backend/requirements.txt
flask==2.0.3
flask-cors==5.0.0
torch==1.10.2
scikit-image==0.19.3
numpy==1.21.6
Pillow==9.0.1
